backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0
slowapi==0.1.9

# Pydantic for data validation
pydantic==2.10.5
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.36
alembic==1.13.1
psycopg[binary]==3.2.12

# Redis and Caching
redis==5.0.1
hiredis==2.3.2

# Celery for async tasks
celery==5.3.6
flower==2.0.1

# Google Gemini AI
google-generativeai==0.3.2

# HTTP Client
httpx==0.26.0

# HTML Parsing and Validation
beautifulsoup4==4.12.3
html5lib==1.1
lxml==5.1.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
pytest-cov==4.1.0
httpx==0.26.0
rich==13.7.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
